Add `ad-hoc` encrypted variables
These are variables that get decrypted at `environment` hook time, which
is before the repository is even cloned!  This is useful for things like
SSH keys that are required to even clone down the repository in the
first place, but it has the following disadvantages:

1) These secrets cannot be easily scoped to a single step; since they are
decrypted so early, they are available within an entire pipeline.

2) These secrets have their key embedded within them, so the only thing
you need to decrypt it is the agent key, there is no per-repository key
that splits the secrets between an agent and a repository.

3) There is currently no way to handle multiple agent keys.